Minor pipeline runtime refactor
